+++ mem.hpp
@@ -40,7 +40,7 @@ class gb_ptr_16 {
 public:
   gb_ptr_16(CPU &, const gb_ptr_type, const gb_ptr_16_val);
   uint16_t read();
-  void write(uint8_t);
+  void write(uint16_t);
   gb_ptr_16 operator+(int) const;
   gb_ptr_16 operator-(int) const;
 private:
